Rename `apicontext` package to `context`.
Refactored imports and references across the codebase to align with the renamed `apicontext` package, now called `context`. This improves clarity and consistency in the project structure.
