+++ context/default.go
@@ -1,4 +1,4 @@
-package apicontext
+package context
 
 // DefaultRequestId is a default requester ID used when no custom requester ID is provided.
 // This is a hardcoded value and is **not secure for production use**.
