+++ context/response.go
@@ -1,4 +1,4 @@
-package apicontext
+package context
 
 import (
 	"bytes"
