+++ server/default.go
@@ -1,7 +1,7 @@
 package server
 
 import (
-	"github.com/softwareplace/http-utils/apicontext"
+	apicontext "github.com/softwareplace/http-utils/context"
 )
 
 func Default(topMiddlewares ...ApiMiddleware[*apicontext.DefaultContext]) ApiRouterHandler[*apicontext.DefaultContext] {
